# Speed up 10x test_storage_rabbitmq::test_failed_connection