--- a/tests/integration/test_storage_rabbitmq/test_failed_connection.py
+++ b/tests/integration/test_storage_rabbitmq/test_failed_connection.py
@@ -172,6 +172,9 @@ def rabbitmq_monitor():
 
 # Tests
 
+timeout = 60
+
+
 def test_rabbitmq_restore_failed_connection_without_losses_1(rabbitmq_cluster, rabbitmq_monitor):
     instance.query(
         """
@@ -182,8 +185,8 @@ def test_rabbitmq_restore_failed_connection_without_losses_1(rabbitmq_cluster, r
         CREATE TABLE test.consume (key UInt64, value UInt64)
             ENGINE = RabbitMQ
             SETTINGS rabbitmq_host_port = 'rabbitmq1:5672',
-                    rabbitmq_flush_interval_ms=500,
-                    rabbitmq_max_block_size = 100,
+                    rabbitmq_flush_interval_ms=1000,
+                    rabbitmq_max_block_size = 1,
                     rabbitmq_exchange_name = 'producer_reconnect',
                     rabbitmq_format = 'JSONEachRow',
                     rabbitmq_num_consumers = 2,
@@ -197,14 +200,15 @@ def test_rabbitmq_restore_failed_connection_without_losses_1(rabbitmq_cluster, r
                     rabbitmq_exchange_name = 'producer_reconnect',
                     rabbitmq_persistent = '1',
                     rabbitmq_flush_interval_ms=1000,
+                    rabbitmq_max_block_size = 1,
                     rabbitmq_format = 'JSONEachRow',
                     rabbitmq_row_delimiter = '\\n';
     """
     )
 
-    messages_num = 300000
+    messages_num = 1000
     rabbitmq_monitor.set_expectations(published=messages_num, delivered=messages_num)
-    deadline = time.monotonic() + 180
+    deadline = time.monotonic() + timeout
     while time.monotonic() < deadline:
         try:
             instance.query(
@@ -218,25 +222,29 @@ def test_rabbitmq_restore_failed_connection_without_losses_1(rabbitmq_cluster, r
                 raise
     else:
         pytest.fail(
-            f"Time limit of 180 seconds reached. The query could not be executed successfully."
+            f"Time limit of {timeout} seconds reached. The query could not be executed successfully."
         )
 
-    deadline = time.monotonic() + 180
+    deadline = time.monotonic() + timeout
     while time.monotonic() < deadline:
         number = int(instance.query("SELECT count() FROM test.view"))
-        logging.debug(f"{number}/{messages_num} before suspending RabbitMQ")
         if number != 0:
-            if number == messages_num:
-                pytest.fail("The RabbitMQ messages have been consumed before suspending the RabbitMQ server")
+            logging.debug(f"{number}/{messages_num} before suspending RabbitMQ")
             break
         time.sleep(0.1)
     else:
-        pytest.fail(f"Time limit of 180 seconds reached. The count is still 0.")
+        pytest.fail(f"Time limit of {timeout} seconds reached. The count is still 0.")
 
     suspend_rabbitmq(rabbitmq_cluster, rabbitmq_monitor)
+
+    number = int(instance.query("SELECT count() FROM test.view"))
+    logging.debug(f"{number}/{messages_num} after suspending RabbitMQ")
+    if number == messages_num:
+        pytest.fail("All RabbitMQ messages have been consumed before resuming the RabbitMQ server")
+
     resume_rabbitmq(rabbitmq_cluster, rabbitmq_monitor)
 
-    deadline = time.monotonic() + 180
+    deadline = time.monotonic() + timeout
     while time.monotonic() < deadline:
         result = instance.query("SELECT count(DISTINCT key) FROM test.view")
         if int(result) == messages_num:
@@ -245,7 +253,7 @@ def test_rabbitmq_restore_failed_connection_without_losses_1(rabbitmq_cluster, r
         time.sleep(1)
     else:
         pytest.fail(
-            f"Time limit of 180 seconds reached. The result did not match the expected value."
+            f"Time limit of {timeout} seconds reached. The result did not match the expected value."
         )
 
     instance.query(
@@ -268,9 +276,9 @@ def test_rabbitmq_restore_failed_connection_without_losses_2(rabbitmq_cluster, r
             ENGINE = RabbitMQ
             SETTINGS rabbitmq_host_port = 'rabbitmq1:5672',
                     rabbitmq_exchange_name = 'consumer_reconnect',
-                    rabbitmq_num_consumers = 10,
-                    rabbitmq_flush_interval_ms = 100,
-                    rabbitmq_max_block_size = 100,
+                    rabbitmq_num_consumers = 2,
+                    rabbitmq_flush_interval_ms = 1000,
+                    rabbitmq_max_block_size = 1,
                     rabbitmq_num_queues = 10,
                     rabbitmq_format = 'JSONEachRow',
                     rabbitmq_row_delimiter = '\\n';
@@ -282,9 +290,9 @@ def test_rabbitmq_restore_failed_connection_without_losses_2(rabbitmq_cluster, r
     """
     )
 
-    messages_num = 300000
+    messages_num = 1000
     rabbitmq_monitor.set_expectations(published=messages_num, delivered=messages_num)
-    deadline = time.monotonic() + 180
+    deadline = time.monotonic() + timeout
     while time.monotonic() < deadline:
         try:
             instance.query(
@@ -298,22 +306,26 @@ def test_rabbitmq_restore_failed_connection_without_losses_2(rabbitmq_cluster, r
                 raise
     else:
         pytest.fail(
-            f"Time limit of 180 seconds reached. The query could not be executed successfully."
+            f"Time limit of {timeout} seconds reached. The query could not be executed successfully."
         )
 
-    deadline = time.monotonic() + 180
+    deadline = time.monotonic() + timeout
     while time.monotonic() < deadline:
         number = int(instance.query("SELECT count() FROM test.view"))
-        logging.debug(f"{number}/{messages_num} before suspending RabbitMQ")
         if number != 0:
-            if number == messages_num:
-                pytest.fail("The RabbitMQ messages have been consumed before suspending the RabbitMQ server")
+            logging.debug(f"{number}/{messages_num} before suspending RabbitMQ")
             break
         time.sleep(0.1)
     else:
-        pytest.fail(f"Time limit of 180 seconds reached. The count is still 0.")
+        pytest.fail(f"Time limit of {timeout} seconds reached. The count is still 0.")
 
     suspend_rabbitmq(rabbitmq_cluster, rabbitmq_monitor)
+
+    number = int(instance.query("SELECT count() FROM test.view"))
+    logging.debug(f"{number}/{messages_num} after suspending RabbitMQ")
+    if number == messages_num:
+        pytest.fail("All RabbitMQ messages have been consumed before resuming the RabbitMQ server")
+
     resume_rabbitmq(rabbitmq_cluster, rabbitmq_monitor)
 
     # while int(instance.query('SELECT count() FROM test.view')) == 0:
@@ -322,7 +334,7 @@ def test_rabbitmq_restore_failed_connection_without_losses_2(rabbitmq_cluster, r
     # kill_rabbitmq()
     # revive_rabbitmq()
 
-    deadline = time.monotonic() + 180
+    deadline = time.monotonic() + timeout
     while time.monotonic() < deadline:
         result = instance.query("SELECT count(DISTINCT key) FROM test.view").strip()
         if int(result) == messages_num:
@@ -331,7 +343,7 @@ def test_rabbitmq_restore_failed_connection_without_losses_2(rabbitmq_cluster, r
         time.sleep(1)
     else:
         pytest.fail(
-            f"Time limit of 180 seconds reached. The result did not match the expected value."
+            f"Time limit of {timeout} seconds reached. The result did not match the expected value."
         )
 
     instance.query(